==> notebooks/1.1 Preprocessing_domain_1.ipynb <==
# Preprocessing the current listing property from the domain website
This notebook is to preprocess the listing property data that were scraped from the script.py.


### Import packages

import pandas as pd
import re
import numpy as np

# Read the Listing Data

listings_df = pd.read_parquet('../data/landing/listings.parquet')

listings_df.head()

# Set a index for each row
listings_df['link'] = listings_df.index
listings_df.index = range(listings_df.shape[0])

# check the size of the data
listings_df.shape

# For coulmns as lists, retrieve the actual values

for i in range(listings_df.shape[0]):
    
    # number of beds
    num_beds_str = listings_df.loc[i, 'rooms']
    if num_beds_str:
        num_beds = re.findall('\d+', num_beds_str[0])
        if num_beds:
            listings_df.loc[i, 'rooms'] = num_beds[0]
        else:
            listings_df.loc[i, 'rooms'] = np.nan
    else:
        listings_df.loc[i, 'rooms'] = np.nan

    
    # number of baths
    num_bath_str = listings_df.loc[i, 'bath']
    if num_bath_str:
        num_bath = re.findall('\d+', num_bath_str[0])
        if num_bath:
            listings_df.loc[i, 'bath'] = num_bath[0]
        else:
            listings_df.loc[i, 'bath'] = np.nan
    else:
        listings_df.loc[i, 'bath'] = np.nan


    # number of parking
    num_parking_str = listings_df.loc[i, 'parking']
    if num_parking_str: 
        num_parking = re.findall('\d+', num_parking_str[0])
        if num_parking:
            listings_df.loc[i, 'parking'] = num_parking[0]
        else:
            listings_df.loc[i, 'parking'] = np.nan        
    else:
        listings_df.loc[i, 'parking'] = np.nan


# Column type changing
# Since int type is not supported for columns with nan, changed to float instead
listings_df['rooms'] = listings_df['rooms'].astype(float)
listings_df['bath'] = listings_df['bath'].astype(float)
listings_df['parking'] = listings_df['parking'].astype(float)

## Explode the school columns

The school column contain three schools in order, the first one being the closest public school, the second one being the second closest public school, the third one being the closest private school, the school distance contain the distance of the three schools in order 

def school_filler(df, i, col, name, dis):
    ''' fill the the distance and name of the ith listing's related school'''
    df.loc[i, col] = name
    df.loc[i, f'{col}_dis'] = dis

for i in range(listings_df.shape[0]):
    schools = listings_df.loc[i, 'school_names']
    distance = listings_df.loc[i, 'school_distance']
    if type(schools)!=type(None):
        try:
            school_filler(listings_df, i, 'public_1', schools[0], re.findall('(.+)\skm\saway', distance[0])[0])
            school_filler(listings_df, i, 'public_2', schools[1], re.findall('(.+)\skm\saway', distance[1])[0])
            school_filler(listings_df, i, 'private_1', schools[2], re.findall('(.+)\skm\saway', distance[2])[0])
        except IndexError:
            '''when an index error happens, the there's no information about the closest 1st/2nd 
            public and 1st private school and we just leave the cell being nan'''
            pass


listings_df.head()

# change the type of distance columns to float
listings_df['public_1_dis'] = listings_df['public_1_dis'].astype(float)
listings_df['public_2_dis'] = listings_df['public_2_dis'].astype(float)
listings_df['private_1_dis'] = listings_df['private_1_dis'].astype(float)

# drop the original column
listings_df.drop(columns = ['school_names', 'school_distance'], inplace = True)

# Percentage Features value extracting

for i in range(listings_df.shape[0]):
    # extract the numbers in the percentages
    
    try:
        listings_df.loc[i, 'LT_resident_pcg'] = re.findall('(.+)\%', listings_df.loc[i, 'LT_resident_pcg'])[0]
    except IndexError:
        listings_df.loc[i, 'LT_resident_pcg'] = np.nan
    except TypeError:
        listings_df.loc[i, 'LT_resident_pcg'] = np.nan

    
    try:
        listings_df.loc[i, 'owner_pcg'] = re.findall('(.+)\%', listings_df.loc[i, 'owner_pcg'])[0]
    except IndexError:
        listings_df.loc[i, 'owner_pcg'] = np.nan
    except TypeError:
        listings_df.loc[i, 'owner_pcg'] = np.nan


    try:
        listings_df.loc[i, 'family_pcg'] = re.findall('(.+)\%', listings_df.loc[i, 'family_pcg'])[0]
    except IndexError:
        listings_df.loc[i, 'family_pcg'] = np.nan   
    except TypeError:
        listings_df.loc[i, 'family_pcg'] = np.nan

# change the type of the percentage columns to float
listings_df['LT_resident_pcg'] = listings_df['LT_resident_pcg'].astype(float)
listings_df['owner_pcg'] = listings_df['owner_pcg'].astype(float)
listings_df['family_pcg'] = listings_df['family_pcg'].astype(float)

listings_df.head(10)

# Extract the weekly price

lease_type = []
for i in range(listings_df.shape[0]):
    # extract the weekly price and lower all the other letters in the price column
    try:
        listings_df.loc[i, 'rent'] = re.findall('\$([\d\,]+)', listings_df.loc[i, 'price'])[0]
        listings_df.loc[i, 'rent'] = listings_df.loc[i, 'rent'].replace(',','')
        listings_df.loc[i, 'price'] = listings_df.loc[i, 'price'].lower()
        if re.findall('[A-z]+',listings_df.loc[i, 'price']) not in lease_type:
            lease_type.append(re.findall('[a-z]+',listings_df.loc[i, 'price']))
    except TypeError:
        pass
    except IndexError:
        pass

# change the type of the column to float
listings_df['rent'] = listings_df['rent'].astype(float)

# number of weeks per month constant 
wpm = 4.35
# number of week per year
wpy = wpm * 12
# number of week per season
wps = wpm * 3
for i in range(listings_df.shape[0]):
    # check if the price is weekly or monthly, weekly is assumed if unspecified
    # It has been checked there is no annualy contracts available
    try:
        text = listings_df.loc[i, 'price']
        if 'month' in text or 'mth' in text or 'pm' in text and 'pw' not in text and 'week' not in text:
            listings_df.loc[i, 'rent'] = listings_df.loc[i, 'rent'] / wpm

        elif 'season' in text:
            listings_df.loc[i, 'rent'] = listings_df.loc[i, 'rent'] / wps

        elif 'p.a' in text:
            listings_df.loc[i, 'rent'] = listings_df.loc[i, 'rent'] / wpy
            
        if ('furnished' in text or 'furniture' in text or 'furn' in text) and 'not' not in text:
            listings_df.loc[i, 'is_furnished'] = True
        else:
            listings_df.loc[i, 'is_furnished'] = False 
    except TypeError:
        pass

listings_df.head()

# save the file to raw
listings_df.to_csv('../data/raw/listings.csv', index = False)
