==> notebooks/1.3 Preprocessing_population_density.ipynb <==
# Preprocess population density data from 2001 to 2022

This notebook is to preprocess the population density data that was downloaded from ABS

### Import package

import pandas as pd
import requests


# Download data and read data

# Download 2021-2022 population data to landing

url = "https://www.abs.gov.au/statistics/people/population/regional-population/2021-22/32180DS0001_2021-22r.xlsx"
response = requests.get(url)

with open('../data/landing/population2021-22.xlsx', 'wb') as file:
    file.write(response.content)

df = pd.read_excel('../data/landing/population2021-22.xlsx',sheet_name="Table 2", engine="openpyxl", skiprows=7)
df.head(10)

# Download 2001-2022 population data to landing

url = "https://www.abs.gov.au/statistics/people/population/regional-population/2021-22/32180DS0003_2001-22.xlsx"
response = requests.get(url)

with open('../data/landing/population2001-22.xlsx', 'wb') as file:
    file.write(response.content)

df = pd.read_excel('../data/landing/population2001-22.xlsx',sheet_name="Table 1", engine="openpyxl", skiprows=7)
df.columns


# Select and save Victoria data in population2001-2022 to raw

df_vic = df[df['S/T name'] == 'Victoria']

# change column names
years = list(range(2001, 2023))
cols = df_vic.columns.tolist()
cols[10:] = years
df_vic.columns = cols

df_vic.to_csv('../data/raw/population2001-22_VIC.csv', index=False)

df_vic.columns


# Calculate population density for 2001-2022

df_population = pd.read_csv('../data/raw/population2001-22_VIC.csv')
df_area = pd.read_csv('../data/raw//SA2_area.csv')

# inner merge datasets on 'SA2 code'
merged_df = pd.merge(df_population, df_area, on='SA2 code', how='inner')

merged_df.columns

# calculate population density for 2001-2022
for year in range(2001,2023):
    col_name = str(year)  # Making sure the column name is a string
    density_col_name = f'{year} Density'
    merged_df[density_col_name] = merged_df[col_name] / merged_df['km2']


merged_df.columns
