==> notebooks/1.4 Preprocessing_population_household.ipynb <==
# Preprocessing 2021 ABS census data on population

This notebook is to preprocess the external population data per SA2 in 2021

## Import package

import pandas as pd

#### Preprocessing Population data
First, we want to know how many people per SA2 code in 2021


# load 2021 ABS Census data on family number
family_count_2021_path = "../data/landing/ABS_data/SA2_(EN)_by_FNOF_Family_Number.csv"
skip_rows = 9
family_count_2021_data = pd.read_csv(family_count_2021_path, skiprows=skip_rows)
family_count_2021_data

family_count_2021_data.info()

family_count_2021_data = family_count_2021_data.applymap(lambda x: 0 if "Cells "
                                                                          "in this table have been randomly adjusted to avoid the release of confidential data. "
                                                                          "No reliance should be "
                                                                          "placed on small cells."
                                                                          "" in str(x) else x)

family_count_2021_data['Primary family'] = family_count_2021_data['Primary family'].astype(float)
family_count_2021_data.info()

# preprocessing
family_count_2021_data.drop(family_count_2021_data.index[0], inplace = True)
family_count_2021_data.drop(columns=family_count_2021_data.columns[-1], inplace = True)
family_count_2021_data

family_count_2021_data.rename(columns={'Not applicable': 'Extended Family'}, inplace=True)
family_count_2021_data.rename(columns={'FNOF Family Number': 'SA2 Name'}, inplace=True)
family_count_2021_data.rename(columns={'Total': 'Total Suburb Population'}, inplace=True)

family_count_2021_data

# family_count_2021_data.info()
family_count_2021_data['Immediate Family'] = family_count_2021_data['Total Suburb Population'] - family_count_2021_data['Extended Family']
family_count_2021_data

family_count_2021_data.drop(columns=family_count_2021_data.columns[1:4], inplace = True)
family_count_2021_data

# rearrange columns positions
family_count_2021_data = family_count_2021_data[['SA2 Name', 'Immediate Family', 'Extended Family', 'Total Suburb Population']]
family_count_2021_data

family_count_2021_data = family_count_2021_data.iloc[:-6, :]
family_count_2021_data.tail(10)

# Specify the file path where you want to save the CSV file
family_file_path = "../data/curated/ABS_data/SA2_Family_Number_2021.csv"

# Use the to_csv method to save the DataFrame to a CSV file
family_count_2021_data.to_csv(family_file_path, index=False)

#### Then, we want to count how many people

# load 2021 ABS Census data on family number
person_count_2021_path = "../data/landing/ABS_data/SA2_(EN)_by_CPRF_Count_of_Persons_in_Family.csv"

skip_rows = 9
person_count_2021_data = pd.read_csv(person_count_2021_path, skiprows=skip_rows)
person_count_2021_data

# preprocessing
person_count_2021_data.drop(person_count_2021_data.index[0], inplace = True)
person_count_2021_data.drop(columns=person_count_2021_data.columns[-1], inplace = True)
person_count_2021_data

person_count_2021_data.info()

person_count_2021_data = person_count_2021_data.applymap(lambda x: 0 if "Cells "
                                                                        "in this table have been randomly adjusted to avoid the release of confidential data. "
                                                                        "No reliance should be "
                                                                        "placed on small cells."
                                                                        "" in str(x) else x)


person_count_2021_data['Two persons in family'] = person_count_2021_data['Two persons in family'].astype(float)
person_count_2021_data.info()

person_count_2021_data.rename(columns={'CPRF Count of Persons in Family': 'SA2 Name'}, inplace=True)
person_count_2021_data.rename(columns={'Total': 'Total Suburb Population'}, inplace=True)

person_count_2021_data


person_count_2021_data = person_count_2021_data.iloc[:-6, :]
person_count_2021_data.tail(10)

#### Save the data

# Specify the file path where you want to save the CSV file
person_file_path = "../data/curated/ABS_data/SA2_Person_in_Family_2021.csv"

# Use the to_csv method to save the DataFrame to a CSV file
person_count_2021_data.to_csv(person_file_path, index=False)
