==> notebooks/1.5 Preprocessing_dwelling.ipynb <==
## Preprocessing Dwelling data in 2021
This notebook is to preprocess dwelling types in victoria per SA2

import pandas as pd

# SA2_(EN)_by_STRD_Dwelling_Structure.csv

# load 2021 ABS Census data on dwelling structure
dwelling_types_2021_path = "../data/landing/ABS_data/SA2_(EN)_by_STRD_Dwelling_Structure.csv"

skip_rows = 9
dwelling_types_2021_data = pd.read_csv(dwelling_types_2021_path, skiprows=skip_rows)
dwelling_types_2021_data

dwelling_types_2021_data.info()

dwelling_types_2021_data = dwelling_types_2021_data.applymap(lambda x: 0 if "Cells "
                                                                          "in this table have been randomly adjusted to avoid the release of confidential data. "
                                                                          "No reliance should be "
                                                                          "placed on small cells."
                                                                          "" in str(x) else x)

dwelling_types_2021_data['Separate house'] = dwelling_types_2021_data['Separate house'].astype(float)
dwelling_types_2021_data.info()

# preprocessing
dwelling_types_2021_data.drop(dwelling_types_2021_data.index[0], inplace = True)
dwelling_types_2021_data.drop(columns=dwelling_types_2021_data.columns[-1], inplace = True)
dwelling_types_2021_data

# SA2 by HCFMD Family Household Composition (Dwelling).csv
dwelling_types_2021_data.rename(columns={'STRD Dwelling Structure': 'SA2 Name'}, inplace=True)
dwelling_types_2021_data.rename(columns={'Total': 'Total Suburb Dwellings'}, inplace=True)

dwelling_types_2021_data

dwelling_types_2021_data['Semi-detached house or Townhouse'] = dwelling_types_2021_data['Semi-detached, row or terrace house, townhouse etc. with one storey'] + dwelling_types_2021_data['Semi-detached, row or terrace house, townhouse etc. with two or more storeys']
dwelling_types_2021_data

dwelling_types_2021_data['Flat or Apartment'] = dwelling_types_2021_data['Flat or apartment attached to a house'] + dwelling_types_2021_data['Flat or apartment in a four to eight storey block'] + dwelling_types_2021_data['Flat or apartment in a nine or more storey block'] + dwelling_types_2021_data['Flat or apartment in a one or two storey block'] + dwelling_types_2021_data['Flat or apartment in a three storey block'] + dwelling_types_2021_data['Flat or apartment attached to a house']

dwelling_types_2021_data['Temporary housing'] = dwelling_types_2021_data['Caravan'] + dwelling_types_2021_data['Cabin, houseboat'] + dwelling_types_2021_data['Improvised home, tent, sleepers out']

dwelling_types_2021_data = dwelling_types_2021_data.drop(columns=['Semi-detached, row or terrace house, townhouse etc. with two or more storeys', 'Semi-detached, row or terrace house, townhouse etc. with one storey'])

dwelling_types_2021_data['Not Applicable'] = dwelling_types_2021_data['Not applicable'] + dwelling_types_2021_data['Not stated']

dwelling_types_2021_data


# dwelling_types_2021_data = dwelling_types_2021_data.drop(columns=['column1', 'column2'])
dwelling_types_2021_data.drop(columns=dwelling_types_2021_data.columns[2:6], inplace = True)
dwelling_types_2021_data

dwelling_types_2021_data.drop(columns=dwelling_types_2021_data.columns[2:6], inplace = True)
dwelling_types_2021_data

dwelling_types_2021_data.drop(columns=dwelling_types_2021_data.columns[3:5], inplace = True)
dwelling_types_2021_data

dwelling_types_2021_data['Non-residential properties'] = dwelling_types_2021_data['Not Applicable'] + dwelling_types_2021_data['Temporary housing'] + dwelling_types_2021_data['House or flat attached to a shop, office, etc.']
dwelling_types_2021_data

dwelling_types_2021_data = dwelling_types_2021_data.drop(columns=['Not Applicable', 'House or flat attached to a shop, office, etc.', 'Temporary housing'])
dwelling_types_2021_data

# rearrange columns positions
dwelling_types_2021_data = dwelling_types_2021_data[['SA2 Name', 'Separate house', 'Semi-detached house or Townhouse', 'Flat or Apartment', 'Non-residential properties', 'Total Suburb Dwellings']]
dwelling_types_2021_data

dwelling_types_2021_data = dwelling_types_2021_data.iloc[:-6, :]
dwelling_types_2021_data.tail(10)

We dropped "Migratory - Offshore - Shipping (Vic.)" because from the ABS (1270.0.55.001 - Australian Statistical Geography Standard (ASGS): Volume 1 - Main Structure and Greater Capital City Statistical Areas, July 2016, https://www.abs.gov.au/ausstats/abs@.nsf/lookup/by%20subject/1270.0.55.001~july%202016~main%20features~statistical%20area%20level%202%20(sa2)~10014)

it mentions that "there are non-spatial SA2s for Migratory - Offshore - Shipping" therefore, it is meaningless to include this entry


##### Save the data

# Specify the file path where you want to save the CSV file
file_path = "../data/curated/ABS_data/SA2_Dwelling_Structure_2021.csv"

# Use the to_csv method to save the DataFrame to a CSV file
dwelling_types_2021_data.to_csv(file_path, index=False)
