==> notebooks/1.6 Preprocessing_income.ipynb <==
# Preprocessing income
This notebook is to preprocess the income data and join to the domain listing.

### Import packages

import pandas as pd

### Read the file

income_path = "../data/landing/ABS_data/income.csv"

income_df = pd.read_csv(income_path)

income_df.head()

income_df.shape

income_df.info()

# extract the weekly income column
income_df['Weekly Income($)'] = income_df['value'].str.replace('$', '').str.replace(',', '').astype(float)
income_df.head()


selected_columns = ['suburb', 'Weekly Income($)']
final_income = income_df[selected_columns]
final_income.head()

## Save the file

# Specify the file path where you want to save the CSV file
file_path = "../data/raw/ABS_data/income_suburb.csv"

# Use the to_csv method to save the DataFrame to a CSV file
final_income.to_csv(file_path, index=False)

## Join with the property data

income_sa2_path = "../data/landing/ABS_data/income_per_person_sa2.csv"
income_sa2 = pd.read_csv(income_sa2_path)
income_sa2.head()

income_sa2 = income_sa2.rename(columns={'SA2 NAME': 'SAL_NAME21'})
income_sa2.head()

final_income.head()

final_path = '../data/raw/final_data_1.csv'
final_df = pd.read_csv(final_path)
final_df.head()

### Drop unnecessary columns

final_df.drop('Weekly Income($)', axis=1, inplace=True)
final_df.head()

final_df.drop('nearest_park_distance_km', axis=1, inplace=True)
final_df.head()

drop_columns = ['LT_resident_pcg', 'owner_pcg', 'family_pcg']
final_df.drop(columns = ['LT_resident_pcg', 'owner_pcg', 'family_pcg'], inplace = True)
final_df.dropna()

final_df.head()

file_path = "../data/raw/final_drop_pcg.csv"

# Use the to_csv method to save the DataFrame to a CSV file
final_df.to_csv(file_path, index=False)

## Save the final datasets for processing

result = pd.merge(final_df, income_sa2, on='SAL_NAME21', how='outer')
result.head()

# Specify the file path where you want to save the CSV file
file_path = "../data/raw/ABS_data/final_2.csv"

# Use the to_csv method to save the DataFrame to a CSV file
result.to_csv(file_path, index=False)
