==> notebooks/1.7 Preprocessing_school.ipynb <==
# Preprocess school data
This notebook is to preprocess the school data. At the end, we will get the total number of schools in each suburb.

### Import packages

import pandas as pd

# read csv
csv_file_path = '../data/raw/schoollocations2022.csv'
school = pd.read_csv(csv_file_path,  encoding='ISO-8859-1')

import geopandas as gpd
from shapely.geometry import Point
import folium


## Read the file

# drop columns
columns_to_drop = ['Entity_Type','SCHOOL_NO', 'School_Status','Postal_Address_Line_1','Postal_Address_Line_2','Postal_Town','Postal_State','Postal_Postcode','Full_Phone_No','LGA_ID','LGA_Name']
school_drop = school.drop(columns=columns_to_drop)

school_drop

# Get the shpaefile for SA2 region
suburbs = gpd.read_file("../data/landing/SAL_2021_AUST_GDA2020_SHP/SAL_2021_AUST_GDA2020.shp")


# select the subset that we care about
suburbs = suburbs.loc[suburbs['STE_NAME21']=='Victoria']
suburbs = suburbs[['SAL_CODE21','SAL_NAME21','geometry']]

# Convert the dataframe to a geodataframe
geometry = [Point(xy) for xy in zip(school_drop['X'], school_drop['Y'])]
school_drop_gdf = gpd.GeoDataFrame(school_drop, geometry=geometry)

# Use sjoin to find out which suburb each point belongs to
joined = gpd.sjoin(school_drop_gdf, suburbs, how="left", op="within")

# If your suburbs shapefile has a column named 'suburb_name' that specifies the name of the suburb
school_drop_gdf['SAL_NAME'] = joined['SAL_NAME21']
school_drop_gdf['SAL_CODE'] = joined['SAL_CODE21']

school_drop_gdf

school_drop_gdf.to_csv("../data/curated/school_suburbs_SA2.csv", index=False)

school_drop_gdf.to_file("../data/curated/school_suburbs_SA2.shp")

# Group the data by 'SA2_NAME21' and count the number of 'School_Name' for each group
SA2_school_count = school_drop_gdf.groupby(['SAL_NAME', 'SAL_CODE'])['School_Name'].count().reset_index()

# Rename the 'School_Name' column to 'Count' for clarity
SA2_school_count = SA2_school_count.rename(columns={'School_Name': 'School_Count'})

SA2_school_count

# Check for duplicate values in 'School_Name'
duplicates = SA2_school_count[SA2_school_count['SAL_NAME'].duplicated(keep=False)]

# If 'duplicates' DataFrame is not empty, it means there are two or more rows with the same value in 'Column1'
if not duplicates.empty:
    print("There is duplicate in SAL_NAME")
else:
    print("There are no duplicate in SAL_NAME")

SA2_school_count.to_csv('../data/curated/School_count_with_SA2.csv',index=False)
