==> notebooks/1.8 Preprocessing_inflation.ipynb <==
# Inflation
This notebooke is to preprocess inflation file


# import packages

# 2021SA2 (UR) by 1-digit level OCCP Occupation.csv
import pandas as pd


# load 2021 ABS Census data on family number
inflation_rate_path = "../data/landing/inflation_rate.csv"
skip_rows = 16
inflation_rate_data = pd.read_csv(inflation_rate_path, skiprows=skip_rows)
inflation_rate_data

inflation_rate_data.drop(columns=inflation_rate_data.columns[-1], inplace = True)
inflation_rate_data

inflation_rate_data.info()

# Specify the file path where you want to save the CSV file
file_path = "../data/curated/inflation.csv"

# Use the to_csv method to save the DataFrame to a CSV file
inflation_rate_data.to_csv(file_path, index=False)
