==> notebooks/1.9 Preprocessing_occupation.ipynb <==
## Preprocessing occupational data from 2021 census

# 2021SA2 (UR) by 1-digit level OCCP Occupation.csv
import pandas as pd

# load 2021 ABS Census data on family number
occupation_2021_path = "../data/landing/ABS_data/2021SA2 (UR) by 1-digit level OCCP Occupation.csv"
skip_rows = 9
occupation_2021_data = pd.read_csv(occupation_2021_path, skiprows=skip_rows)
occupation_2021_data

occupation_2021_data.info()

occupation_2021_data = occupation_2021_data.applymap(lambda x: 0 if "Cells "
                                                                          "in this table have been randomly adjusted to avoid the release of confidential data. "
                                                                          "No reliance should be "
                                                                          "placed on small cells."
                                                                          "" in str(x) else x)

occupation_2021_data['Managers'] = occupation_2021_data['Managers'].astype(float)
occupation_2021_data.info()

# preprocessing
occupation_2021_data.drop(occupation_2021_data.index[0], inplace = True)
occupation_2021_data.drop(columns=occupation_2021_data.columns[-1], inplace = True)
occupation_2021_data


occupation_2021_data.rename(columns={'1-digit level OCCP Occupation': 'SA2 Name'}, inplace=True)
occupation_2021_data.rename(columns={'Total': 'Total Suburb Occupation Types'}, inplace=True)


occupation_2021_data = occupation_2021_data.iloc[:-7, :]
occupation_2021_data.tail(10)


#### Save the data

# Specify the file path where you want to save the CSV file
file_path = "../data/curated/ABS_data/SA2_Occupation_Types_2021.csv"

# Use the to_csv method to save the DataFrame to a CSV file
occupation_2021_data.to_csv(file_path, index=False)
