==> notebooks/1.10 Preprocessing_postcode.ipynb <==
# Preprocessing Postcode data

This notebook we will download and preprocess the postcode data 

import pandas as pd

## Download the postcode data to local landing folder

%run ../scripts/Postcode_Download.py

postcode_df = pd.read_csv('../data/landing/postcode/australian_postcodes.csv')
print(postcode_df.columns)

### Feature selection

# Define a list of column names to retain
columns_to_retain = ['postcode', 'locality', 'state', 'SA2_MAINCODE_2016', 'SA2_NAME_2016', 'Lat_precise', 'Long_precise']  # Replace with your desired column names

# Filter the DataFrame to retain only the specified columns
postcode_df = postcode_df[columns_to_retain]

postcode_df

### Select Victoria region

vic_postcode = postcode_df[postcode_df['state'] == 'VIC']
# now we don't need state columns anymore
vic_postcode = vic_postcode.drop(columns= 'state')

vic_postcode

aware that there is some data containing 0 Lat_precise	and Long_precis.

Here we assume that any row of data contain NaN is invalid and thus need ot be remove

# remove all row of data that has nan
vic_postcode = vic_postcode.dropna()
vic_postcode

### Save the files

# Save to raw data 
vic_postcode.to_csv('../data/raw/location_postcode.csv',index=False)
