==> notebooks/1.11 Preprocessing_park.ipynb <==
# Park data
This notebook is to preprocess the park data in vic

### Read files 

import pandas as pd

csv_file_path = '../data/raw/location_postcode(1).csv'
postcode = pd.read_csv(csv_file_path).dropna().drop_duplicates(subset=['SA2_NAME_2016'])




csv_file_path = '../data/raw/parks-and-reserves1.csv'
park = pd.read_csv(csv_file_path)  

# print(df.head())


### Get the suburb file

postcode['SA2_NAME_2016'] = [x.replace(" ", "").upper().split('-')[0].strip() for x in postcode['SA2_NAME_2016'].tolist() if x is not None]
postcode

postcode.dtypes

Park preprocessing

### Join with the park dataset

park

# drop columns
columns_to_drop = ['geo_point_2d', 'postcode', 'ward_name','easting','northing','projection','googlemaps_drive_to','area']
park_drop = park.drop(columns=columns_to_drop) # 创建一个新的 DataFrame，不修改原始 DataFrame
#park_drop = park_drop.drop(park_drop.columns[0], axis=1)
park_drop

# drop null
park_drop = park_drop.dropna(subset=['name'])
filtered_park = park_drop[park_drop['name'].str.contains('garden|reserve', case=False, regex=True)]


filtered_park['suburb_name'] = [x.replace(" ", "").strip() for x in filtered_park['suburb_name'].tolist()]

filtered_park.count()

filtered_park.columns

filtered_park.dtypes

postcode.columns = ['postcode', 'locality', 'SA2_MAINCODE_2016',
                    'suburb_name', 'latitude', 'longitude']
postcode = postcode.loc[:, ['postcode', 'suburb_name']].drop_duplicates().drop_duplicates(subset=['suburb_name'])
postcode

merge_df = pd.merge(filtered_park, postcode, on=['suburb_name'], how='left')

merge_df

merge_df[merge_df['postcode'].isnull()]

### Save the files

merge_df.to_csv('../data/curated/finally_park.csv', index=False)
