==> notebooks/1.12 Preprocessing_SA2.ipynb <==
# Preprocess SA2 shapefile
This notebook is to preprocess the SA2 shapefile, retriving relevant columns for geovisualisation in the later steps


#### Import packages

import pandas as pd

pip install openpyxl

#### Read the file

# load 2021 ABS Census data on family number
# table 2
SA2_listing_2021_path = "../data/raw/2021-22 SA2 ERP VIC.xlsx"
skip_rows = 7
SA2_listing_2021_data = pd.read_excel(SA2_listing_2021_path, sheet_name='Table 2', skiprows=skip_rows)
SA2_listing_2021_data

SA2_listing_2021_data.drop(SA2_listing_2021_data.index[0], inplace = True)
SA2_listing_2021_data

selected_columns = [['Name', 'City']]
