==> notebooks/1.13 Preprocessing_shopping_center.ipynb <==
# Preprocess shopping center data
This notebook is to preprocess the shopping center data, at the end, we will get the number of shopping centers in each suburb and their geo location

### Import packages

import re
import pandas as pd
import numpy as np
import geopandas as gpd
from shapely.geometry import Point

### Read the files

shopping_centers = pd.read_csv('../data/landing/shopping_center.csv')

shopping_centers.head()

# Retrieve information for each shopping center 

def fill_shopping_info (i, state, city, lantitude, longitude, num_store):
    '''fill in the ith row of a dataframe with relevant information about a shopping center'''
    shopping_centers.loc[i, 'state'] = state
    shopping_centers.loc[i, 'city'] = city 
    shopping_centers.loc[i, 'longitude'] = longitude
    shopping_centers.loc[i, 'latitude'] = lantitude
    shopping_centers.loc[i, 'num_store'] = num_store

for i in range(shopping_centers.shape[0]):
    info = shopping_centers.loc[i, 'information']
    # match the informations
    try:
        retrieved_info= re.findall('(.+),\s(.+),\sGPS:\s(.+),\s(.+)\s\|\sphone:\s(?:\+61\s)?.+\s\|\sstores:\s(.+)', info)[0]
        fill_shopping_info(i, *retrieved_info)

    # some shopping contain information in a different format, we deal with them seperately
    except IndexError:
        if i in [30,174,189,198]:
            retrieved_info = re.findall('(.+),\s(.+),\sGPS:\s(.+),\s(.+)\s\|\sphone:.+', info)[0]
            fill_shopping_info(i, *retrieved_info, num_store=np.nan)

        else:
            retrieved_info = re.findall('(.+),\s(.+),\sGPS:\s(.+),\s(.+)\s\s\|\sstores:\s(.+)', info)[0]
            fill_shopping_info(i, *retrieved_info)           

shopping_centers.drop('information', inplace=True, axis=1)
shopping_centers.head()

# number of store should be an integer however, since it contain nan, we leave it as float
shopping_centers['num_store'] = shopping_centers['num_store'].astype(float)

# the longitude and lantitude info should be float instead of string
shopping_centers['longitude'] = shopping_centers['longitude'].astype(float)
shopping_centers['latitude'] = shopping_centers['latitude'].astype(float)

shopping_centers.dtypes

# Get the SA2 region for each shopping center

# read the shpae file that contain SA2 info
suburbs = gpd.read_file("../data/curated/SA2_2021_AUST_GDA2020.shp")

# get the relevant columns
suburbs = suburbs.loc[suburbs['STE_NAME21']=='Victoria']
suburbs = suburbs[['SA2_CODE21','SA2_NAME21','geometry']]

# Convert the shopping center dataframe to a gdf
geometry = [Point(xy) for xy in zip(shopping_centers['longitude'], shopping_centers['latitude'])]
shopping_gdf = gpd.GeoDataFrame(shopping_centers, geometry=geometry)

# Use sjoin to find out which suburb each point belongs to
joined = gpd.sjoin(shopping_gdf, suburbs, how="left", op="within")

# If your suburbs shapefile has a column named 'suburb_name' that specifies the name of the suburb
shopping_gdf['SA2_NAME'] = joined['SA2_NAME21']
shopping_gdf['SA2_CODE'] = joined['SA2_CODE21']

# saved the shopping centers to curated data file
shopping_gdf.to_csv('../data/curated/shopping_centers.csv', index=False)
