==> notebooks/1.14 Preprocessing_crime.ipynb <==
# Preprocess Crime data

This notebook will run the `Crime_Download.py` script in the scripts folder to download the crime data to local landing folder

This notebook will then continue preprocessing the data, grouping/counting the crime instance record by location/postcode/time(year)

### Import packages

import pandas as pd


import geopandas as gpd
from fuzzywuzzy import fuzz

import geopandas as gpd
from shapely.geometry import Point

## Dowload Crime data to landing folder

%run ../scripts/Crime_Download.py

## Read in our desire table/sheet of crime data 
as table 3 store content of Criminal incidents by principal offence, local government area and postcode or suburb/town - April 2013 to March 2023

# Read the "Table 03" sheet into a DataFrame
sheet_name = "Table 03"
crime_df = pd.read_excel("../data/landing/crime_stat/crime_stat_2023.xlsx", sheet_name=sheet_name)

crime_df

## Feature selection

# feature selection
crime_df = crime_df[['Year', 'Postcode', 'Suburb/Town Name', 'Incidents Recorded']]
crime_df

# Apply agrregation

#### Group by 'Suburb/Town Name', 'Postcode', 'Year' and count the total crime cases by suburbs

# Group by 'Suburb/Town Name', 'Postcode', 'Year'
suburb_crime_counts = crime_df.groupby(['Suburb/Town Name', 'Postcode', 'Year'])['Incidents Recorded'].sum().reset_index()

suburb_crime_counts

# Save to Raw
suburb_crime_counts.to_csv('../data/raw/suburb_crime_counts.csv', index=False)  # Set index=False to exclude the index column

However, it is important to note that the suburb/Town Name may not be consistent with our SAL name as data is from different provider.
We will find a point geometry for each suburb for later visualisation and merging

# Create a point geometry to the suburb_crime_counts

To do this we will consider combining the crime count data with victoria' postcode data frame.

As postcode data contain postcode (a common feature) and different locality (suburb) which we will be use to match with the "Suburb/Town Name" attribute in our crime data frame by similarity score. 

Most importantly postcode data contains point geometry for the locality.

## Combine crime count with victoria postcode data 
Join by postcode, and select for most similar name for the locality name and Suburb/Town Name from two the data 

# import the postcode data
Postcode_df = pd.read_csv("../data/raw/location_postcode.csv")

Postcode_df

Notice how the locality value are all in capital case, we will convert all Suburb/Town Name in the suburb_crime_counts and locality in postcode to lowercase for laters string comparision 

We will not using SA2_NAME and CODE in postcode df as they were from 2016, we will mainly focusing on the locality (suburb) and the lat and long for laters merging 

# Lower case the Suburb/Town Name in crime count data frame
suburb_crime_counts['Suburb/Town Name'] = suburb_crime_counts['Suburb/Town Name'].str.lower()
# Lower case the locality in postcode data
Postcode_df['locality'] = Postcode_df['locality'].str.lower()
# Drop SA2_MAINCODE_2016 and SA2_NAME_2016 in postcode data
Postcode_df = Postcode_df.drop(columns=['SA2_MAINCODE_2016', 'SA2_NAME_2016'])

#### First join suburb_crime_counts and  Postcode_df by left join on postcode number

# Merge the crime data with postcode by postcode number first
crime_count_n_locality= suburb_crime_counts.merge(Postcode_df, left_on='Postcode', right_on='postcode', how='left')

crime_count_n_locality

# Check if there is Suburb/Town Name that didn't get a macth
crime_count_n_locality[crime_count_n_locality.isnull().any(axis=1)]

#### Then compute similarity score between locality and Suburb/Town Name, and select the matched data/row with highest simialrity score for each (Suburb/Town Name)

# Create a new column name similarityscore that store similarlity score between locality and Suburb/Town Name
crime_count_n_locality['SimilarityScore'] = crime_count_n_locality.apply(
    lambda row: fuzz.ratio(row['Suburb/Town Name'], row['locality']),
    axis=1
)

crime_count_n_locality

#### Select the matched locality with highest Simialrity Score with Suburb/Town Name	

# Group by 'Suburb/Town Name' and 'Year' and select row of the data with the highest 'SimilarityScore'
crime_count_n_locality = crime_count_n_locality.loc[crime_count_n_locality.groupby(['Suburb/Town Name','Year'])['SimilarityScore'].idxmax()]


crime_count_n_locality

#### Check data with poor Simialrity score

# Check data with mis_matched Suburb/Town Name to locality
mis_matched = crime_count_n_locality[crime_count_n_locality['SimilarityScore'] < 100]
mis_matched

By obseriving the data frame we see that there are 18 row of data has mismatched Suburb/Town Name and locality. Which suggest that most data were correcly matched between (Yeahhhhhhhhhhhh).

We can choose to retain these data as they all coming from same postcode region, and their lat and long will be somewhere close to their actual location.

However, as these mis matched data mostly coming from "cokum", ""watchem west", "timbarra", "eddington", "lake eppalock" and "mirimbah" with each only contain 1-2 incident record. Hence we can choose to drop these data out as they will be negligible 

# Retain rows with correctly matched Suburb/Town Name === 'SimilarityScore'=100
crime_count_n_locality = crime_count_n_locality[crime_count_n_locality['SimilarityScore'] == 100]
crime_count_n_locality.reset_index(drop=True, inplace=True)


crime_count_n_locality

### Feature Selection/Drop duplicate/unnecessary columns 

# Drop the specified columns
crime_count_df = crime_count_n_locality.drop(columns=['postcode', 'locality', 'SimilarityScore'])


crime_count_df

#### Convert the crime_count_n_locality to a geodataframe 

# Convert the dataframe to a geodataframe
geometry = [Point(xy) for xy in zip(crime_count_df['Long_precise'], crime_count_df['Lat_precise'])]
crime_count_gpd = gpd.GeoDataFrame(crime_count_df, geometry=geometry)
# drop lat and long columns
crime_count_gpd = crime_count_gpd.drop(columns=['Long_precise', 'Lat_precise'])


crime_count_gpd

### Save as a copy in raw

crime_count_gpd.to_file('../data/raw/crime_count_with_point_geo.geojson', driver="GeoJSON")  # Specify the driver based on the desired output format
