==> notebooks/1.15 Preprocessing_public_transport.ipynb <==
# PTV data
This notebook is to download ptv data, which included the public transport information about trains, trams, buses in Victoria state. Then we will preprocess the data and use the raw file for our analysis.

Important Notice! 

The url for PTV data is temporary generate and share by Victoria Datashare (https://datashare.maps.vic.gov.au/). If The link is expired, please follow the instruction in README.MD file to reorder the dataset


Please Run SA2_&_SAL_shapefile_preprocessing.ipynb first before run this notebook 

import os 
import fiona
import geopandas as gpd
import pandas as pd

## Run the download script to download ptv data file 


url = "https://s3.ap-southeast-2.amazonaws.com/cl-isd-prd-datashare-s3-delivery/Order_HI08BK.zip" 

%run ../scripts/PTV_Download.py


## Exaimine the datafile

ptv_layers = fiona.listlayers('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb')
ptv_layers

# Extract Train Data 

# load and preprocess train data
## including: 1. metro train, 2.regional train

gdf_REGIONAL_TRAIN= gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', 
layer='PTV_REGIONAL_TRAIN_STATION')

gdf_METRO_TRAIN = gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', 
layer='PTV_METRO_TRAIN_STATION')

df_mtrain_stops = pd.DataFrame({'TRAIN_STOP_ID': gdf_METRO_TRAIN['STOP_ID'], 'STOP_NAME': gdf_METRO_TRAIN['STOP_NAME'], 
'geometry': gdf_METRO_TRAIN['geometry'], 'type': pd.Series(["METRO_TRAIN" for x in range(len(gdf_METRO_TRAIN.index))])})

df_rtrain_stops = pd.DataFrame({'TRAIN_STOP_ID': gdf_REGIONAL_TRAIN['STOP_ID'], 'STOP_NAME': gdf_REGIONAL_TRAIN['STOP_NAME'],
'geometry': gdf_REGIONAL_TRAIN['geometry'], 'type': pd.Series(["REGIONAL_TRAIN" for x in range(len(gdf_REGIONAL_TRAIN.index))])})

df_train_stops = pd.concat([df_mtrain_stops, df_rtrain_stops], axis=0)

# Save to landing
df_train_stops.to_csv('../data/landing/PTV DATA/train_stops.csv',index=False)

## Match the public transport station location with the SAL region/suburb based on the geometry location (we define SAL region to be suburb)

## Import SAL data 

file_path = "../data/raw/victoria_region_gdf/SAL_region_gdf.geojson"
# Read the GeoPandas DataFrame from the specified file
SAL_gpd = gpd.read_file(file_path)

## Merge the Train stop data with SAL data by geometry join

#convert the train_data to geopandas
train_stops_gdf = gpd.GeoDataFrame(
    df_train_stops,
    geometry=df_train_stops['geometry']
)

# join by geometry join
train_stops_with_SAL = gpd.sjoin(train_stops_gdf, SAL_gpd, how='left', op='within')

train_stops_with_SAL

##### Check if there is unmatched data

train_stops_with_SAL[train_stops_with_SAL['SAL_NAME21'].isna()]

Notice this is stop in outside victoria, hence we will remove it 

## Remove rows where "SA2_NAME21" is None

train_stops_with_SAL = train_stops_with_SAL.dropna(subset=['SAL_NAME21'])


train_stops_with_SAL[train_stops_with_SAL['SAL_NAME21'].isna()]

### Feature selection

# remove unnecessary columns
train_stops_with_SAL = train_stops_with_SAL.drop(columns=['STOP_NAME', 'geometry', 'index_right', 'SHAPE_Area'])

train_stops_with_SAL

# Rename the train type column
train_stops_with_SAL = train_stops_with_SAL.rename(columns={'type': 'train_type'})

# Save the this to raw data 

# make directory in raw data file to store filtered ptv data
output_relative_dir = '../data/raw/PTV'

if not os.path.exists(output_relative_dir):
    os.makedirs(output_relative_dir)
    

train_stops_with_SAL.to_csv('../data/raw/PTV/train_stops_with_SAL.csv',index=False)

# Similalry with Bus 

# load and preprocess bus data
## including: 1. regional bus, 2. metro bus, 3. skybus, 4. regional coach

gdf_REGIONAL_BUS= gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', 
layer='PTV_REGIONAL_BUS_STOP')

gdf_METRO_BUS= gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', 
layer='PTV_METRO_BUS_STOP')

gdf_SKYBUS= gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', 
layer='PTV_SKYBUS_STOP')

gdf_REGIONAL_COACH= gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', 
layer='PTV_REGIONAL_COACH_STOP')

df_REGIONAL_BUS_stops = pd.DataFrame({'BUS_STOP_ID': gdf_REGIONAL_BUS['STOP_ID'], 'STOP_NAME': gdf_REGIONAL_BUS['STOP_NAME'], 
'geometry': gdf_REGIONAL_BUS['geometry'], 'type': pd.Series(["REGIONAL_BUS" for x in range(len(gdf_REGIONAL_BUS.index))])})

df_METRO_BUS = pd.DataFrame({'BUS_STOP_ID': gdf_METRO_BUS['STOP_ID'], 'STOP_NAME': gdf_METRO_BUS['STOP_NAME'], 
'geometry': gdf_METRO_BUS['geometry'], 'type': pd.Series(["METRO_BUS" for x in range(len(gdf_METRO_BUS.index))])})

df_SKYBUS = pd.DataFrame({'BUS_STOP_ID': gdf_SKYBUS['STOP_ID'], 'STOP_NAME': gdf_SKYBUS['STOP_NAME'], 
'geometry': gdf_SKYBUS['geometry'], 'type': pd.Series(["SKYBUS" for x in range(len(gdf_SKYBUS.index))])})

df_REGIONAL_COACH = pd.DataFrame({'BUS_STOP_ID': gdf_REGIONAL_COACH['STOP_ID'], 'STOP_NAME': gdf_REGIONAL_COACH['STOP_NAME'],
'geometry': gdf_REGIONAL_COACH['geometry'], 'type': pd.Series(["REGIONAL_COACH" for x in range(len(gdf_REGIONAL_COACH.index))])})

df_bus_stops = pd.concat([df_REGIONAL_BUS_stops, df_METRO_BUS, df_SKYBUS, df_REGIONAL_COACH], axis=0)
# Save to landing PTV
df_bus_stops.to_csv('../data/landing/PTV DATA/bus_stops.csv',index=False)

df_bus_stops

## Merge the Bus data with SAL region based on their geometry location 

#### first convert the df_bus_stops to geopandas df

#convert the train_data to geopandas
bus_stops_gdf = gpd.GeoDataFrame(
    df_bus_stops,
    geometry=df_bus_stops['geometry']
)

Merge bus_stops_gdf with our victoria region shapefile 

# merge bus stop data with SAL by geometry join
bus_stops_with_SAL = gpd.sjoin(bus_stops_gdf, SAL_gpd, how='left', op='within')

#### Check if there is unmatched data

bus_stops_with_SAL[bus_stops_with_SAL['SAL_NAME21'].isna()]

notice how these unmatched bus data were all regional bus with bus stop outside victoria state. 

we will simply disgard these data, assuming that these bus stop has no effect on housing renting in victoria

## Remove rows where "SAL_NAME21" is None
bus_stops_with_SAL = bus_stops_with_SAL.dropna(subset=['SAL_NAME21'])

bus_stops_with_SAL[bus_stops_with_SAL['SAL_NAME21'].isna()]

### Feature selection on bus stop data

# remove unnecessary columns
bus_stops_with_SAL = bus_stops_with_SAL.drop(columns=['STOP_NAME', 'geometry', 'index_right', 'SHAPE_Area'])

# Rename the bus type column
bus_stops_with_SAL = bus_stops_with_SAL.rename(columns={'type': 'bus_type'})

bus_stops_with_SAL

# # Save the this in raw/PTV folder with other PTV datas
bus_stops_with_SAL.to_csv('../data/raw/PTV/bus_stops_with_SAL.csv',index=False)

# Tram

# load and preprocess tram data
gdf_METRO_TRAM = gpd.read_file('../data/landing/PTV DATA/ll_gda94/filegdb/whole_of_dataset/victoria/PTV.gdb', layer='PTV_METRO_TRAM_STOP')

df_tram_stops = pd.DataFrame(gdf_METRO_TRAM)

df_tram_stops.rename(columns={'STOP_ID': 'TRAM_STOP_ID'}, inplace=True)

# Save to landing/PTV data
df_tram_stops.to_csv('../data/landing/PTV DATA/tram_stops.csv',index=False)

df_tram_stops

#### Check for duplicate values in 'Tram_STOP_ID'

# Check for duplicate values in 'BUS_STOP_ID'
duplicates = df_tram_stops[df_tram_stops['TRAM_STOP_ID'].duplicated(keep=False)]

# If 'duplicates' DataFrame is not empty, it means there are two or more rows with the same value in 'TRAM_STOP_ID'
if not duplicates.empty:
    print("There are duplicate value in TRAM_STOP_ID")
    print(duplicates)
else:
    print("There are no duplicate value in TRAM_STOP_ID")

### Merge the Bus stop data with SAL data
first convert the df_tram_stops to geopandas df

#convert the train_data to geopandas
tram_stops_gdf = gpd.GeoDataFrame(
    df_tram_stops,
    geometry=df_tram_stops['geometry']
)

Merge with the victoria region shapefile 

# merge tram data with SAL by geomery location
tram_stops_with_SAL = gpd.sjoin(tram_stops_gdf, SAL_gpd, how='left', op='within')

tram_stops_with_SAL

#### Check if there is unmatched tram data 

tram_stops_with_SAL[tram_stops_with_SAL['SAL_NAME21'].isna()]

#### Apply feature selection on tram stop data 

# remove unnecessary columns
tram_stops_with_SAL = tram_stops_with_SAL.drop(columns=['LATITUDE', 'LONGITUDE', 'STOP_NAME', 'geometry', 'index_right', 'SHAPE_Area', 'TICKETZONE', 'ROUTES_USING_STOP'])

tram_stops_with_SAL

# Save the this to raw/PTV folder
tram_stops_with_SAL.to_csv('../data/raw/PTV/tram_stops_with_SAL.csv',index=False)
