==> notebooks/1.17 Preprocessing_suburb_shapefile.ipynb <==
# SA2 and SAL geo data preprocessing 

This notebook will download and preprocess Victoria's SA2 and SAL geo data file and save individual SA2 and SAL geo data file as csv in folder victoria_region_gdf created in raw data folder 

import os
import zipfile
import pandas as pd
import geopandas as gpd
import re




## Download SA2 and SAL data to local landing folder 

#

## Read and Extract SAL and SA2 zip to landing 
### Create a new folder name shapefile in landing to store extracted data SAL and SA2



# Specify the path to your zip file
SA2_zip_file_path = '../data/landing/SA2_2021_AUST_SHP_GDA2020.zip'

SAL_zip_file_path = '../data/landing/SAL_2021_AUST_GDA2020_SHP.zip'

# Specify the directory where you want to extract the files
SA2_extraction_path = '../data/landing/shapefile/SA2'

SAL_extraction_path = '../data/landing/shapefile/SAL'


# Create the extraction directory if it doesn't exist 
os.makedirs(SA2_extraction_path, exist_ok=True)
os.makedirs(SAL_extraction_path, exist_ok=True)



# Open the SAL zip file
with zipfile.ZipFile(SAL_zip_file_path, 'r') as zip_ref:
    # Extract all the contents to the extraction directory
    zip_ref.extractall(SAL_extraction_path)


# Open the SA2 zip file
with zipfile.ZipFile(SA2_zip_file_path, 'r') as zip_ref:
    # Extract all the contents to the extraction directory
    zip_ref.extractall(SA2_extraction_path)

## Read shapefiles in Geopandas

# Read the shapefile using GeoPandas
### SA2
SA2_gdf = gpd.read_file("../data/landing/shapefile/SA2/SA2_2021_AUST_GDA2020.shp")

### SAL
SAL_gdf = gpd.read_file("../data/landing/shapefile/SAL/SAL_2021_AUST_GDA2020_SHP/SAL_2021_AUST_GDA2020.shp")

SAL_gdf

SA2_gdf

# Select region to victoria
victoria_SA2_gdf = SA2_gdf[SA2_gdf['STE_NAME21']=='Victoria']

victoria_SAL_gdf = SAL_gdf[SAL_gdf['STE_NAME21']=='Victoria']

### Feature selection

# Select key features 
victoria_SA2_gdf = victoria_SA2_gdf.loc[:, ['SA2_NAME21', 'SA2_CODE21', 'geometry']]
victoria_SAL_gdf = victoria_SAL_gdf.loc[:, ['SAL_NAME21', 'SAL_CODE21', 'SHAPE_Area', 'geometry']]


victoria_SA2_gdf

victoria_SAL_gdf

### Remove unnecessary symbols in the string value (Cleaning)

Aware that SA2_NAME21 and SAL_NAME21 contain some, we would like to drop them for later string comparision

# Define a list of symbols to remove
symbols_to_remove = ['-', '_', '(Vic.)', '(', ')', '/', ]

# Remove symbols from all columns containing string values in SA2 name 
victoria_SA2_gdf["SA2_NAME21"] = victoria_SA2_gdf["SA2_NAME21"].str.replace('|'.join(map(re.escape, symbols_to_remove)), '', regex=True)


victoria_SAL_gdf["SAL_NAME21"] = victoria_SAL_gdf["SAL_NAME21"].str.replace('|'.join(map(re.escape, symbols_to_remove)), '', regex=True)

victoria_SA2_gdf

victoria_SAL_gdf

### Drop regions with no geometry 

# check rows with no geometry
display(victoria_SA2_gdf[victoria_SA2_gdf['geometry']==None])
display(victoria_SAL_gdf[victoria_SAL_gdf['geometry']==None])

## Drop these columns 
victoria_SA2_gdf=victoria_SA2_gdf.dropna()
victoria_SAL_gdf=victoria_SAL_gdf.dropna()

## Check
display(victoria_SA2_gdf[victoria_SA2_gdf['geometry']==None])
display(victoria_SAL_gdf[victoria_SAL_gdf['geometry']==None])

### Save the SA2 and SAL shapefile to raw data 

We will first create a new folder in ../data/raw name victoria_region_gdf 

## Create victoria_region_gdf in raw data

# Specify the output directory and filename
output_directory = "../data/raw"
SA2_output_filename = "SA2_region_gdf.geojson" 
SAL_output_filename = "SAL_region_gdf.geojson" 

new_folder_name = "victoria_region_gdf"

# Create the new folder
new_folder_path = os.path.join(output_directory, new_folder_name)
if not os.path.exists(new_folder_path):
    os.makedirs(new_folder_path)
    print(f"New folder created at: {new_folder_path}")
else:
    print(f"Folder already exists at: {new_folder_path}")

# Save the GeoPandas DataFrame to ../data/raw/victoria_region_gdf
SA2_output_path = os.path.join(new_folder_path, SA2_output_filename)
SAL_output_path = os.path.join(new_folder_path, SAL_output_filename)

# save as GeoJson file 
victoria_SA2_gdf.to_file(SA2_output_path, driver="GeoJSON")  
victoria_SAL_gdf.to_file(SAL_output_path, driver="GeoJSON") 


#### Check by reading 

#Check reading it 
SA2_file_path = "../data/raw/victoria_region_gdf/SA2_region_gdf.geojson"

SAL_file_path = "../data/raw/victoria_region_gdf/SAL_region_gdf.geojson"


# Read the GeoPandas DataFrame
SA2_region_gdf = gpd.read_file(SA2_file_path)

SAL_region_gdf = gpd.read_file(SAL_file_path)

SAL_region_gdf

SA2_region_gdf
