==> notebooks/2.1 Analysing_ABS.ipynb <==
# Analysis external ABS datasets
- This notebook is to analysis the external data that were downloaded from the ABS, and merged to the listing domain data

External datasets are:
- population density
- population counts
- income







### Read the shapefiles

from shapely.geometry import Point
import geopandas as gpd
import pandas as pd

shapefile_path = "../data/landing/SA2_2021_AUST_SHP_GDA2020/SA2_2021_AUST_GDA2020.shp"

listing_path = "../data/curated/listings_suburbs_SA2.csv"

suburbs_path = "../data/landing/SAL_2021_AUST_GDA2020_SHP/SAL_2021_AUST_GDA2020.shp"

SA2_list = gpd.read_file(shapefile_path)

listing_domain = pd.read_csv(listing_path)
listing_domain

### Retrieve relevant columns

SA2_list = SA2_list.loc[SA2_list['STE_NAME21']=='Victoria']
SA2_list = SA2_list[['SA2_NAME21','AREASQKM21']]
SA2_list.head()

suburb_list = gpd.read_file(suburbs_path)
suburb_list.head()

suburb_list = SA2_list[['SAL_NAME21','AREASQKM21']]
suburb_list.head()

#### Read the saved ABS files

# income_file_path = "../data/landing/ABS_data/income.csv"
dwelling_file_path = "../data/curated/ABS_data/SA2_Dwelling_Structure_2021.csv"
occupation_file_path = "../data/curated/ABS_data/SA2_Occupation_Types_2021.csv"
person_file_path = "../data/curated/ABS_data/SA2_Person_in_Family_2021.csv"
family_file_path = "../data/curated/ABS_data/SA2_Family_Number_2021.csv"

occupation_data = pd.read_csv(occupation_file_path)
occupation_data.head()

dwelling_data = pd.read_csv(dwelling_file_path)
dwelling_data.head()

person_data = pd.read_csv(person_file_path)
person_data.head()

family_data = pd.read_csv(family_file_path)
family_data.head()

merged_df = pd.merge(family_data, person_data, on="SA2 Name", how="inner")
merged_df = pd.merge(merged_df, dwelling_data, on="SA2 Name", how="inner")

merged_df.head()

columns_to_drop = ["Total Suburb Population_y", "Not applicable"]
merged_df.drop(columns=columns_to_drop, inplace=True)

merged_df.head()

merged_df.rename(columns={"Total Suburb Population_x": "Total Suburb Population"}, inplace=True)

merged_df.head()

### merged to domain listing

merged_df.head()

merged_df.rename(columns={"SA2 Name": "SA2_NAME"}, inplace=True)
merged_df.head()

listing_domain.head()

final_dataset = pd.merge(listing_domain, merged_df, on="SA2_NAME", how="inner")
final_dataset.head()

file_path = "../data/curated/ABS_data/ABS_domain.csv"

# Use the to_csv method to save the DataFrame to a CSV file
final_dataset.to_csv(file_path, index=False)
