==> notebooks/2.3 Analysing_park.ipynb <==
# Park data
This notebook is to analyse the park data for each suburb.

### Import packages

import geopandas as gpd
from shapely.geometry import Point
import folium
import pandas as pd

### Read files

# read files
csv_file_path = '../data/raw/parks-and-reserves1.csv'
park = pd.read_csv(csv_file_path)

### Cleaning the data

# drop unnecessary columns
columns_to_drop = ['geo_point_2d', 'postcode', 'ward_name','easting','northing','projection','googlemaps_drive_to','area']
park_drop = park.drop(columns=columns_to_drop)

# drop null
park_drop = park_drop.dropna(subset=['name'])
filtered_park = park_drop[park_drop['name'].str.contains('garden|reserve', case=False, regex=True)]


filtered_park

### Read shapefile to map the park attribute to suburb

# Get the shpaefile for SA2 region
suburbs = gpd.read_file("../data/landing/SAL_2021_AUST_GDA2020_SHP/SAL_2021_AUST_GDA2020.shp")


# select the subset that we care about
suburbs = suburbs.loc[suburbs['STE_NAME21']=='Victoria']
suburbs = suburbs[['SAL_CODE21','SAL_NAME21','geometry']]

suburbs

# Convert the dataframe to a geodataframe
geometry = [Point(xy) for xy in zip(filtered_park['longitude'], filtered_park['latitude'])]
filtered_park_gdf = gpd.GeoDataFrame(filtered_park, geometry=geometry)

# Use sjoin to find out which suburb each point belongs to
joined = gpd.sjoin(filtered_park_gdf, suburbs, how="left", op="within")

# If your suburbs shapefile has a column named 'suburb_name' that specifies the name of the suburb
filtered_park_gdf['SAL_NAME'] = joined['SAL_NAME21']
filtered_park_gdf['SAL_CODE'] = joined['SAL_CODE21']

filtered_park_gdf

### Save the data

filtered_park_gdf.to_csv("../data/curated/park_suburbs_SA2.csv", index=False)

filtered_park_gdf.to_file("../data/curated/park_suburbs_SA2.shp")

# Group the data by 'SA2_NAME21' and count the number of 'park_Name' for each group
SA2_park_count = filtered_park_gdf.groupby(['SAL_NAME', 'SAL_CODE'])['name'].count().reset_index()

# Rename the 'School_Name' column to 'Count' for clarity
SA2_park_count = SA2_park_count.rename(columns={'name': 'Park_Count'})
SA2_park_count

# Check for duplicate values in 'School_Name'
duplicates = SA2_park_count[SA2_park_count['SAL_NAME'].duplicated(keep=False)]

# If 'duplicates' DataFrame is not empty, it means there are two or more rows with the same value in 'Column1'
if not duplicates.empty:
    print("There is duplicate in SAL_NAME")
else:
    print("There are no duplicate in SAL_NAME")

SA2_park_count.to_csv('../data/curated/Park_count_with_SA2.csv',index=False)

### Merge with the listing property data

SA2_school_count = pd.read_csv("../data/curated/School_count_with_SA2.csv")
listings_suburbs = pd.read_csv("../data/curated/PTV_count_with_SA2_with_PTV_count.csv")
listings_suburbs 

SA2_school_count
SA2_park_count

print(SA2_school_count['SAL_NAME'].dtype)
print(SA2_school_count['SAL_CODE'].dtype)
print(SA2_park_count['SAL_NAME'].dtype)
print(SA2_park_count['SAL_CODE'].dtype)


import numpy as np
SA2_park_count['SAL_CODE'] = SA2_park_count['SAL_CODE'].astype(np.int64)

aaa=SA2_park_count[SA2_park_count['SAL_NAME']=='Botanic Ridge']['Park_Count']
aaa

# Merge the dataframes on 'SA2_NAME21' with an outer join
SA2_school_park_count = SA2_school_count.merge(SA2_park_count, on=['SAL_NAME', 'SAL_CODE'], how='outer')

# Fill NaN values with 0 for tram and bus counts
SA2_school_park_count = SA2_school_park_count.fillna(0)
SA2_school_park_count['Park_Count']

print(listings_suburbs['SAL_CODE'].dtype)
print(SA2_school_count['SAL_CODE'].dtype)


#listings_suburbs['SAL_CODE'] = listings_suburbs['SAL_CODE'].astype(str)
#SA2_school_park_count['SAL_CODE'] = SA2_school_park_count['SAL_CODE'].astype(str)


### Save the final dataset

# join based on SAL_NAME21 in both df
listings_suburbs = listings_suburbs .merge(SA2_school_park_count, left_on='SAL_CODE', right_on='SAL_CODE', how='left')

listings_suburbs

listings_suburbs.to_csv('../data/curated/SCHOOL&PARK_count_with_SA2_with_SCHOOL&PARK_count.csv',index=False)
