==> notebooks/2.5 Preliminary_Visualisation_Crime_SAL.ipynb <==
# Crime data With SAL

This notebook will focuse in preliminary visualization of crime data in each SAL region

Before run this, you need to run 1.17 Preprocessing_suburb_shapefile.ipynb and 1.14 Preprocessing_crime.ipynb

import pandas as pd
import geopandas as gpd
from fuzzywuzzy import fuzz
import folium
import numpy as np

# Import data 

# Read crime count data 
crime_count_gpd = gpd.read_file('../data/raw/crime_count_with_point_geo.geojson')  # Specify the driver based on the desired output format
# Read SA2 gdf
SAL_gdf = gpd.read_file("../data/raw/victoria_region_gdf/SAL_region_gdf.geojson")

crime_count_gpd

SAL_gdf

## Now join with victoria SAL file by their geometry

# Use sjoin to find out which suburb each point belongs to
crime_count_gpd = gpd.sjoin(crime_count_gpd, SAL_gdf, how="left", op="within")

crime_count_gpd

#### Retain key features

crime_count_gpd = crime_count_gpd[['SAL_NAME21', 'SAL_CODE21', 'Year', 'Incidents Recorded']]

### Rename the incident recorded to number of crime incident
crime_count_gpd = crime_count_gpd.rename(columns={'Incidents Recorded': 'Crime_incidents_count'})

crime_count_gpd

## Check if one location/suburb have been mapped to same SAL region 

## Check if one location/suburb have been mapped to same SAL region 
temp = crime_count_gpd[['SAL_NAME21', 'SAL_CODE21', 'Year']]

duplicates = temp[temp.duplicated()]
duplicates

We aware this duplication could due to the geographical scale and region for SAL, which could result single SAL region contain one or more suburb, hence we will group by SAL_NAME21	SAL_CODE21	Year and sum up the Crime_incidents_count


# Group by 'SAL_NAME21', 'SAL_CODE21', and 'Year' and sum 'Crime_incidents_count'
crime_count_gpd = crime_count_gpd.groupby(['SAL_NAME21', 'SAL_CODE21', 'Year'])['Crime_incidents_count'].sum().reset_index()

# Rename the 'Crime_incidents_count' column to 'total_crime_incidents'
crime_count_gpd = crime_count_gpd.rename(columns={'Crime_incidents_count': 'total_crime_incidents'})


crime_count_gpd

#### Check number of SAL we cover 

yearly_count = crime_count_gpd.groupby('Year').size().reset_index(name='RecordCount')
yearly_count

good !

# Save to Cureated 
crime_count_gpd.to_csv("../data/curated/crime_count_SAL.csv", index=False)


# Visualisation of Crime data in 2023 with SAL region


# Read data 
crime_count_gpd = pd.read_csv("../data/curated/crime_count_SAL.csv")

# Select the releveant year 2023
crime_2023 = crime_count_gpd[crime_count_gpd['Year']==2023]

#reading  the geo data 
file_path = "../data/raw/victoria_region_gdf/SAL_region_gdf.geojson"
# Read the GeoPandas DataFrame from the specified file
SAL_gdf = gpd.read_file(file_path)

SAL_gdf

crime_2023

crime_2023 =crime_2023.drop(columns=['Year'])
crime_2023

# create a JSON 
geoJSON = SAL_gdf[['SAL_CODE21', 'geometry']].to_json()

# print the first 300 chars of the json
print(geoJSON[:300])

# Draw geo visualisation

m = folium.Map(location=[-36.9848, 143.3906], tiles="Stamen Terrain", zoom_start=7)

# refer to the folium documentations on more information on how to plot aggregated data.
c = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=crime_2023.reset_index(), # data source
    columns=['SAL_CODE21','total_crime_incidents'], # the columns required
    key_on='properties.SAL_CODE21', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color='black',
    legend_name='crime cases'
)

c.add_to(m)

m.save('../plots/preliminary_crime_incidents_SAL.html')



# file size too large, cant show m

# apply log scale to total_crime_incidents and repeat the visualisation

crime_2023['log_total_crime_incidents'] = np.log(crime_2023['total_crime_incidents'])


crime_2023

## Re draw with loged crime incident

k = folium.Map(location=[-36.9848, 143.3906], tiles="Stamen Terrain", zoom_start=7)

# refer to the folium documentations on more information on how to plot aggregated data.
c = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=crime_2023.reset_index(), # data source
    columns=['SAL_CODE21','log_total_crime_incidents'], # the columns required
    key_on='properties.SAL_CODE21', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
   # nan_fill_color='black',
    legend_name='crime cases'
)

c.add_to(k)

k.save('../plots/log_crime_incidents_SAL.html')

# file size too big to show map
#k


# save the data table to curated 
crime_2023.to_csv('../data/curated/preliminary_2023crime_count_under_SAL.csv', index=False)  # Specify the driver based on the desired output format
