==> notebooks/2.6 Analysing_crime_rate.ipynb <==
# Crime rate per SAL region

This notebook will aim to first join the crime data with SAL data.

Then determine crime rate (number of crime instance per population) within each SAL/suburb region 

import pandas as pd
import geopandas as gpd

# Read crime count data 
crime_count_gpd = gpd.read_file('../data/raw/crime_count_with_point_geo.geojson')  # Specify the driver based on the desired output format
# Read SA2 gdf
SAL_gdf = gpd.read_file("../data/raw/victoria_region_gdf/SAL_region_gdf.geojson")
# remove null
SAL_gdf = SAL_gdf.dropna()

### Join the crime data and SAL by geometry join

# Use sjoin to find out which suburb each point belongs to
crime_count_gpd = gpd.sjoin(crime_count_gpd, SAL_gdf, how="left", op="within")

crime_count_gpd

### Feature selection

crime_count_gpd = crime_count_gpd[['SAL_NAME21', 'SAL_CODE21', 'Year', 'Incidents Recorded']]

### Rename the incident recorded to number of crime incident
crime_count_gpd = crime_count_gpd.rename(columns={'Incidents Recorded': 'Crime_incidents_count'})

crime_count_gpd

#### Check if one location/suburb have been mapped to same SAL region 

## Check if one location/suburb have been mapped to same SAL region 
## We temp as a temporary df
temp = crime_count_gpd[['SAL_NAME21', 'SAL_CODE21', 'Year']]

duplicates = temp[temp.duplicated()]
duplicates

We aware this duplication could due to the geographical scale and region for SAL, which could result single SAL region contain one or more locality in the crime data, hence we will group by SAL_NAME21	SAL_CODE21	Year and sum up the Crime_incidents_count

crime_count_gpd

# Group by 'SAL_NAME21', 'SAL_CODE21', and 'Year' and sum 'Crime_incidents_count'
crime_count_gpd = crime_count_gpd.groupby(['SAL_NAME21', 'SAL_CODE21', 'Year'])['Crime_incidents_count'].sum().reset_index()

# Rename the 'Crime_incidents_count' column to 'total_crime_incidents'
crime_count_gpd = crime_count_gpd.rename(columns={'Crime_incidents_count': 'total_crime_incidents'})


# save
crime_count_gpd.to_csv('../data/curated/crime_count_SAL.csv',index=False)

crime_count_gpd

# Select 2023 year

# Select the releveant year 2023
crime_2023 = crime_count_gpd[crime_count_gpd['Year']==2023]


crime_2023['SAL_CODE21'] = crime_2023['SAL_CODE21'].astype(int)


crime_2023

# Import population data

domain_df = pd.read_csv('../data/curated/final_drop_pcg.csv')
# feature selection
domain_df = domain_df[['SAL_NAME21', 'SAL_CODE', 'Total Suburb Population']]
# find population per sal region
population_sal = domain_df.groupby(['SAL_NAME21', 'SAL_CODE'])['Total Suburb Population'].mean().reset_index()


# convert the value type to int
population_sal['SAL_CODE'] = population_sal['SAL_CODE'].astype(int)


population_sal

## merge population and crime data for crime rate

Ensure use left join here, as we also want to consider SAL regions that we have population data but not crime data. We will estimate them by victoria's overall crime rate by its population 

merged = population_sal.merge(crime_2023, left_on='SAL_CODE', right_on='SAL_CODE21', how='left')


merged

# select desire column/features 
merged = merged[['SAL_NAME21_x', 'SAL_CODE', 'Total Suburb Population', 'total_crime_incidents']]


merged 

# check suburb with missing crime data 
merged[merged.isnull().any(axis=1)]

## Estimate missing data by victoria's crime rate multiply with the population

According to research currently victoria population is 6.78million 
the total record crime instance is 371,691 from https://www.crimestatistics.vic.gov.au/media-centre/news/key-figures-year-ending-june-2023

crime rate = crime instance/population


we will estimate any missing value of crime instance by vic crime rate *population

# we will estimate any missing value of crime instance by vic crime rate *population 
vic_population = 6780000
crime_rate = 371691/vic_population
crime_rate

# Estimate total_crime_incidents for rows with missing values
merged['total_crime_incidents'].fillna(merged['Total Suburb Population'] * crime_rate, inplace=True)


merged

# check if there is any null
nan_row_count = merged.isna().any(axis=1).sum()

print("Number of rows with NaN values:", nan_row_count)

# Find crime rate for each suburb  

# create a new column for crime rate
merged['crime_rate'] = merged['total_crime_incidents'] / merged['Total Suburb Population']


merged

# rename the column
merged.rename(columns={'SAL_NAME21_x': 'SAL_NAME21'}, inplace=True)


merged

However, here we only have 753 data as we only have 753 data of SAL with population. 

Here we will estimate the remainning suburb/SAL region's crime rate by victoria's crime rate in 2023 we obtain previosly 0.054821681415929205

Full_SAL = SAL_gdf[['SAL_NAME21', 'SAL_CODE21']]
# convert the value type to int
Full_SAL['SAL_CODE21'] = Full_SAL['SAL_CODE21'].astype(int)


# Join the crime data with all SAL region
merged_full = Full_SAL.merge(merged, left_on='SAL_CODE21', right_on='SAL_CODE', how='left')
merged_full = merged_full[['SAL_NAME21_x', 'SAL_CODE21', 'crime_rate']]

merged_full

# fill null by estimate opf victoria crime rate 
merged_full['crime_rate'].fillna(crime_rate, inplace=True)

merged_full

# Save to Cureated 
merged_full.to_csv("../data/curated/crime_rate.csv", index=False)
