==> notebooks/2.8 Calculating_distance.ipynb <==
# Calculate distance
This notebook is used to calculate distance by cars between property and amenities (eg. public transport stations/nearest shopping center/schools/parks/Melbourne CBD)

This was achieved by self-hosting the web api to local host in docker.

First step was to download the australia open route map to be used for open route service.

Link to the map:
wget http://download.geofabrik.de/australia-oceania/australia-latest.osm.pbf

import pandas as pd
import numpy as np
import json
import requests
from ast import literal_eval
import time
import re
import sys
import os

import math

# read the files
domain_path = "../data/curated/listings_suburbs_SA2.csv"

domain_abs_path = "../data/curated/ABS_data/ABS_domain.csv"

domain_abs = pd.read_csv(domain_abs_path)
domain_abs.head()

domain_abs['coordinate'] = domain_abs[['latitude', 'longitude']].apply(lambda x: [x['latitude'], x['longitude']], axis=1)
domain_abs.head()

## Calculate distance to Melbourne CBD

#### Define a function to calculate distance to CBD

# osrm_url = "http://localhost:5000/ors/v2/directions"

# url to the local host
osrm_url = "http://localhost:8080/ors"

def cbd_distances(property_geo):
    '''
    This function is to calculate distance between two locations in km
    :param property_geo:
    :return distance between each property to CBD in km:
    '''
    # CBD coordinates
    cbd_geo = [-37.815207, 144.963937]

    journey = requests.get(f"{osrm_url}{property_geo[1]},{property_geo[0]};{cbd_geo[1]},{cbd_geo[0]}?steps=true&alternatives=true")


    # extract distance from trip information nested dictionary
    distance_meters = json.loads(journey.content)['routes'][0]['legs'][0]['distance']

    return distance_meters / 1000

#### Define a function to get a list of distance for the domain property data

def calculate_cbd_distance(domain_data):
    '''
    This function to get a list of distance from the domain property data
    :param domain_data:
    :return the distance between each coordinate:
    '''
    cbd_distances = []
    for i in domain_data['coordinate']:
        try:
            cbd_distances.append(cbd_distances(literal_eval(i)))
        except:
            cbd_distances.append(np.nan)
            continue

    return cbd_distances

cbd_distances = calculate_cbd_distance(domain_abs)

domain_abs.loc[:,'CBD_Distance'] = cbd_distances

### Define a function to calculate distance in km

# define the function to calculate the distance between two locations in km
def haversine_distance(lat1, lon1, lat2, lon2):
    '''
    This function is to calculate the distance in km
    :param lat1: latitude for the first location
    :param lon1: longitude for the first location
    :param lat2: latitude for the second location
    :param lon2: longitude for the second location
    :return the distance:
    '''
    radius = 6371.0
    lat1 = math.radians(lat1)
    lon1 = math.radians(lon1)
    lat2 = math.radians(lat2)
    lon2 = math.radians(lon2)

    # calculate distance between geometory locations
    dlon = lon2 - lon1
    dlat = lat2 - lat1

    # define a haversine function
    a = math.sin(dlat / 2)**2 + math.cos(lat1) * math.cos(lat2) * math.sin(dlon / 2)**2
    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))

    # calculation distance
    distance = radius * c

    return distance


#### calculate distance to melbourence cbd

melbourne_cbd = (-37.815207, 144.963937)

# calculate the distance to melbourne CBD
domain_abs['distance_to_melbourne_cbd_km'] = domain_abs['coordinate'].apply(lambda coord: haversine_distance(coord[0], coord[1], melbourne_cbd[0], melbourne_cbd[1]))


domain_abs.head()

#### calculate km to shopping center

shopping_center_path = "../data/raw/shopping_centers.csv"

# read the shopping center file
shopping_center = pd.read_csv(shopping_center_path)
shopping_center.head()

# preprocess the geo location for the shopping center
shopping_center['coordinate'] = shopping_center[['latitude', 'longitude']].apply(lambda x: [x['latitude'], x['longitude']], axis=1)
shopping_center.head()

def find_nearest_shopping_center(property_coord):
    '''
    this function is to get the distance to nearest shopping center
    :param property_coord:
    :return nearest distance to shopping center in km:
    '''
    min_distance = float('inf')
    for shop_coord in shopping_center['coordinate']:
        distance = haversine_distance(property_coord[0], property_coord[1], shop_coord[0], shop_coord[1])
        min_distance = min(min_distance, distance)
    return min_distance



# added the distance to the new column
domain_abs['nearest_shopping_center_distance_km'] = domain_abs['coordinate'].apply(find_nearest_shopping_center)

# print
domain_abs.head()

### calculate to nearest train station

import pandas as pd
import re

# load dataset
train_station_path = "../data/raw/train_stops.csv"
train_station = pd.read_csv(train_station_path)
train_station.head()

# define a function to extract coordinates
def extract_coordinates(point_str):
    '''
    this function is to extract coordinates
    :param point_str:
    :return longitude and latitude:
    '''
    # regex to search
    match = re.search(r'\((-?\d+\.\d+)\s+(-?\d+\.\d+)\)', point_str)
    if match:
        longitude = float(match.group(2))
        latitude = float(match.group(1))
        return (longitude, latitude)
    else:
        return None


# add the geo location to a new column
train_station['extracted_coordinates'] = train_station['geometry'].apply(extract_coordinates)

# print
train_station.head()


def find_nearest_train_statin(property_coord):
    '''
    this function is to find the nearest distance to the train station
    :param property_coord:
    :return nearest distance :
    '''
    min_distance = float('inf')
    for train_coord in train_station['extracted_coordinates']:
        distance = haversine_distance(property_coord[0], property_coord[1], train_coord[0], train_coord[1])
        min_distance = min(min_distance, distance)
    return min_distance

# add the distance to the new column
domain_abs['nearest_train_station_distance_km'] = domain_abs['coordinate'].apply(find_nearest_train_statin)

# print
domain_abs.head()

### Calculate distance to tram stop

# load the data
tram_station_path = "../data/raw/tram_stops.csv"
tram_station = pd.read_csv(tram_station_path)
tram_station.head()

# get the coordinate
tram_station['coordinate'] = tram_station[['LATITUDE', 'LONGITUDE']].apply(lambda x: [x['LATITUDE'], x['LONGITUDE']], axis=1)
tram_station.head()

def find_nearest_tram_statin(property_coord):
    '''
    This function is to calculate the distance to the nearest tram station
    :param property_coord:
    :return nearest distance to the tram station in km:
    '''
    min_distance = float('inf')
    for tram_coord in tram_station['coordinate']:
        distance = haversine_distance(property_coord[0], property_coord[1], tram_coord[0], tram_coord[1])
        min_distance = min(min_distance, distance)
    return min_distance

# add the distance to the new column
domain_abs['nearest_tram_station_distance_km'] = domain_abs['coordinate'].apply(find_nearest_tram_statin)

# print
domain_abs.head()

### Calculate distance to bus stop

# load the data
bus_station_path = "../data/raw/bus_stops.csv"
bus_station = pd.read_csv(bus_station_path)
bus_station.head()

# add the distance to the new column

bus_station['extracted_coordinates'] = bus_station['geometry'].apply(extract_coordinates)

# print
bus_station.head()

def find_nearest_bus_stop(property_coord):
    '''
    this function is to find the nearest distance to the bus stop
    :param property_coord:
    :return the nearest distance in km:
    '''
    min_distance = float('inf')
    for bus_coord in bus_station['extracted_coordinates']:
        distance = haversine_distance(property_coord[0], property_coord[1], bus_coord[0], bus_coord[1])
        min_distance = min(min_distance, distance)
    return min_distance


# add the distance to the new column
domain_abs['nearest_bus_stop_distance_km'] = domain_abs['coordinate'].apply(find_nearest_bus_stop)

# print
domain_abs.head()

### calculate distance to park

# load the dataset
park_path = "../data/raw/finally_park.csv"
park_df = pd.read_csv(park_path)
park_df.head()

# get the coodinate
park_df['coordinate'] = park_df[['latitude', 'longitude']].apply(lambda x: [x['latitude'], x['longitude']], axis=1)
park_df.head()

def find_nearest_park(property_coord):
    '''
    this function is to get the distance to the nearest park
    :param property_coord:
    :return minimal distance in km:
    '''
    min_distance = float('inf')
    for park_coord in park_df['coordinate']:
        distance = haversine_distance(property_coord[0], property_coord[1], park_coord[0], park_coord[1])
        min_distance = min(min_distance, distance)
    return min_distance

# add the distance to the new column
domain_abs['nearest_park_distance_km'] = domain_abs['coordinate'].apply(find_nearest_park)

# print
domain_abs.head()

print(domain_abs)

#### Calculate distance to schools

# load the dataset
school_path = "../data/raw/finally_school.csv"
school_df = pd.read_csv(school_path)
school_df.head()

# get the coordinate
school_df['coordinate'] = school_df[['Y', 'X']].apply(lambda x: [x['Y'], x['X']], axis=1)
school_df.head()

school_df.to_csv('../data/raw/school_geo.csv')

def find_nearest_school(property_coord):
    '''
    this function is to get the distance to the nearest school
    :param property_coord:
    :return minimal distance in km:
    '''
    min_distance = float('inf')
    for school_coord in school_df['coordinate']:
        distance = haversine_distance(property_coord[0], property_coord[1], school_coord[0], school_coord[1])
        min_distance = min(min_distance, distance)
    return min_distance


# add the distance to the new column
domain_abs['nearest_school_distance_km'] = domain_abs['coordinate'].apply(find_nearest_school)

# print
domain_abs.head()

#### Save the file

# domain_abs.to_csv('../data/raw/tmp_with_cbd_dist.csv')

domain_abs.to_csv('../data/raw/property_to_dist.csv')

### Join with income

# load the income data
income_path = "../data/raw/ABS_data/income_suburb.csv"
income_df = pd.read_csv(income_path)
income_df.head()

income_df = income_df.rename(columns={'suburb': 'SAL_NAME21'})
income_df.head()

final_df = pd.merge(domain_abs, income_df, on='SAL_NAME21', how='outer')
final_df.head()

### Save the file


final_df.to_csv('../data/raw/final_data_1.csv')
