==> notebooks/3.1 Visualising_public_transport.ipynb <==
# Public transport visualisation 

This notebook will generate a visualisation of public transport station count within each suburb.

This will include total number of train stops, bus stop and tram stops in each suburb


import pandas as pd
import geopandas as gpd
import folium

## read SAL geometry region file

# Read SAL gdf
SAL_gdf = gpd.read_file("../data/raw/victoria_region_gdf/SAL_region_gdf.geojson")
# remove null
SAL_gdf = SAL_gdf.dropna()

# Find all SAL regions name and id in victoria 
Full_SAL = SAL_gdf[['SAL_NAME21', 'SAL_CODE21']]
# convert the value type to int
Full_SAL['SAL_CODE21'] = Full_SAL['SAL_CODE21'].astype(int)


## Import the data for public transport (station) count within each SAL/suburb 

For this project, as our target is rental market within Victoria regions, we believe that regional transportation has negligible effect on the rental market for individual suburb. Hence, here we only consider total public transport count excluding regionals

#read PTV_count_with_SAL
PTV_count = pd.read_csv('../data/curated/PTV_count_with_SAL.csv')

# Select feature of interest 
PTV_count = PTV_count[['SAL_NAME21', 'SAL_CODE21', 'Total_Public_Transport_Count_No_Regional']]

PTV_count['SAL_CODE21'] = PTV_count['SAL_CODE21'].astype(int)


PTV_count

While some suburb appears to have 0 public transport which seems unreasonable, we believe that nowaday every suburb should be able to have an access to at least 1 public transport. Hence, we will add 1 to every location's public transport count 

## Merge the PTV count with every SAL region, including those missing in ptv data 

# Join by SAL_CODE21
merged_full = Full_SAL.merge(PTV_count, left_on='SAL_CODE21', right_on='SAL_CODE21', how='left')

# fill null value of Total_Public_Transport_Count by 0 
merged_full['Total_Public_Transport_Count_No_Regional'].fillna(0, inplace=True)


# add extra 1 count to every SAL location# Hence we will add 1 count for every SAL location
merged_full['Total_Public_Transport_Count_No_Regional'] = merged_full['Total_Public_Transport_Count_No_Regional'] + 1

# rename column name for SAL_NAME21_x
merged_full.rename(columns={'SAL_NAME21_x': 'SAL_NAME21'}, inplace=True)


# Select desire feature 
merged_full = merged_full[['SAL_NAME21', 'SAL_CODE21', 'Total_Public_Transport_Count_No_Regional']]
# Rename the column for public transport count
merged_full.rename(columns={'Total_Public_Transport_Count_No_Regional': 'Total_Public_Transport_Count'}, inplace=True)

merged_full

# Visualise it !

# create a JSON 
geoJSON = SAL_gdf[['SAL_CODE21', 'geometry']].to_json()
# print the first 300 chars of the json
print(geoJSON[:300])

m = folium.Map(location=[-36.9848, 143.3906], tiles="Stamen Terrain", zoom_start=7)

# refer to the folium documentations on more information on how to plot aggregated data.
c = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=merged_full.reset_index(), # data source
    columns=['SAL_CODE21','Total_Public_Transport_Count'], # the columns required
    key_on='properties.SAL_CODE21', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color='black',
    legend_name='Public Transport Station Count'
)

c.add_to(m)
# I will not display this as it will make this file toooooo big to push up to git 
#m


## Label the Top 5 suburub with most public transport 

sorted_df = merged_full.sort_values(by='Total_Public_Transport_Count', ascending=False)
sorted_df.head(10)

### Create centroid for each region

# (y, x) since we want (lat, long)
SAL_gdf['centroid'] = SAL_gdf['geometry'].apply(lambda x: (x.centroid.y, x.centroid.x))
SAL_gdf[['SAL_NAME21', 'SAL_CODE21', 'centroid']].head()

# Assuming you have a GeoPandas DataFrame called gdf
values_to_find = ['22161', '22391', '22750', '20947', '20224', '22121', '21640', '21203', '21013', '20878']

# Use boolean indexing to filter rows where column 'x' contains the values from the list
filtered_rows = SAL_gdf[SAL_gdf['SAL_CODE21'].isin(values_to_find)]
filtered_rows

for SAL_NAME21, coord in filtered_rows[['SAL_NAME21', 'centroid']].values:
    folium.Marker(
        location=coord, 
        tooltip=folium.Tooltip(SAL_NAME21, permanent=True, direction="below")
    ).add_to(m)

# file will be too large to display the map 
#m

# Save to plot 
m.save('../plots/Public_transport_count_SAL.html')
