==> notebooks/3.2 Visualising_school.ipynb <==
# Visualisation of school count in each suburb 

This notebook will generate visualisation of number of school in each suburb

import pandas as pd
import geopandas as gpd
import folium

### Read in the data

# read the listing file that contain school information
listing = pd.read_csv('../data/raw/SCHOOL&PARK_count_with_SA2_with_SCHOOL&PARK_count.csv')

listing.columns

school_count = listing[['SAL_NAME21', 'SAL_CODE', 'School_Count']]

# drop duplicate 
school_count = school_count.drop_duplicates() 

# replace nan column with 0
school_count = school_count.fillna(0)


# rename column
school_count.rename(columns = {"SAL_CODE": "SAL_CODE21"}, inplace = True)

# Path to the SAL files
SAL_file_path = "../data/raw/victoria_region_gdf/SAL_region_gdf.geojson"

# read the SAL file
SAL_gdf = gpd.read_file(SAL_file_path)

# change the type of the file
SAL_gdf['SAL_CODE21'] = SAL_gdf['SAL_CODE21'].astype('int')

# create a JSON object about location and their respective geomtry
geoJSON = SAL_gdf[['SAL_CODE21', 'geometry']].to_json()

# fill in 0 for the suburbs not in the list, if not in the list, 0 school
school_count = pd.merge(school_count, SAL_gdf, on = 'SAL_CODE21', how = 'right')

# feature selection
sal_school_count = school_count[['SAL_NAME21_y', 'SAL_CODE21', 'School_Count']].fillna(0)
sal_school_count

# the top 5 suburbs with the most school
top5 = sal_school_count.sort_values(by = 'School_Count', ascending = False)[0:5]
top5

# change the type of the code column
SAL_gdf['SAL_CODE21'] = SAL_gdf['SAL_CODE21'].astype('int')

# inner join to get the shape of all the regions in top 5 
top5 = pd.merge(top5, SAL_gdf, on = 'SAL_CODE21', how = 'inner')
top5

# get the centroid for labelling
top5['centroid'] = top5['geometry'].apply(lambda x: (x.centroid.y, x.centroid.x))

# plot the price map for number of school
m1 = folium.Map(location=[-37.8136, 144.9631], tiles="Stamen Terrain", zoom_start=10)

folium.Choropleth(
    geo_data=geoJSON,
    data=sal_school_count,
    columns = ['SAL_CODE21','School_Count'],
    key_on='properties.SAL_CODE21',  
    fill_color='YlOrRd', 
    fill_opacity=0.7,
    line_opacity=0.2,
    legend_name='number of school',
).add_to(m1)


for zone_name, coord in top5[['SAL_NAME21_y', 'centroid']].values:
    folium.Marker(
        location=coord, 
        tooltip=folium.Tooltip(zone_name, permanent=True, direction="below")
    ).add_to(m1)

# File size will be too large if display the map
#m1

m1.save('../plots/School_per_SAL_region.html')
