==> notebooks/3.4 Geovisualisation_rent.ipynb <==
import pandas as pd
import geopandas as gpd
from shapely.geometry import Point
import folium
import numpy as np

# Read the Relevant Datasets

# read the listing data file
listings = gpd.read_file('../data/curated/listings_suburbs_SA2.shp')

# read the shape file for SAL regions
suburbs = gpd.read_file("../data/landing/SAL_2021_AUST_GDA2020_SHP/SAL_2021_AUST_GDA2020.shp")

# read the shape file for SA2 regions
SA2 = gpd.read_file("../data/landing/SA2_2021_AUST_SHP_GDA2020/SA2_2021_AUST_GDA2020.shp")

# select the subset that we care about for the SA2
SA2 = SA2.loc[SA2['STE_NAME21']=='Victoria']
SA2 = SA2[['SA2_CODE21', 'SA2_NAME21', 'geometry']]

# select the subset that we care about for the SAL 
suburbs = suburbs.loc[suburbs['STE_NAME21']=='Victoria']
suburbs = suburbs[['SAL_CODE21', 'SAL_NAME21', 'geometry']]

## plot the median rental price for each SA2 region

# median rent price for each region
rent_price = pd.DataFrame(listings.groupby('SA2_NAME').median()['rent'])

# take the log of rent price for a better plotting
rent_price['rent'] = np.log(rent_price['rent'])

# set the index
rent_price['SA2'] = rent_price.index
rent_price.index = range(rent_price.shape[0])

rent_price_merged = SA2.merge(rent_price, left_on = "SA2_NAME21", right_on = "SA2", how ='inner')

# create a JSON object about location and their respective geomtry
geoJSON = SA2[['SA2_NAME21', 'geometry']].to_json()

# plot the price map for residence
m1 = folium.Map(location=[-37.8136, 144.9631], tiles="Stamen Terrain", zoom_start=10)

folium.Choropleth(
    geo_data=geoJSON,
    data=rent_price,
    columns = ['SA2','rent'],
    key_on='properties.SA2_NAME21',  
    fill_color='YlOrRd', 
    fill_opacity=0.7,
    line_opacity=0.2,
    legend_name='log of median rent price',
).add_to(m1)

m1.save('../plots/rent_price_SA2.html')

# fill the studios that does not have a room feature to default 1
listings['rooms'].fillna(1, inplace=True)

# make a new column that calculate the mean rent price per room
listings['per_room'] = np.log(listings['rent']/listings['rooms'])

# median rent price for each region
rent_price = pd.DataFrame(listings.groupby('SA2_NAME').median()['per_room'])

# set the index
rent_price['SA2'] = rent_price.index
rent_price.index = range(rent_price.shape[0])

rent_price_merged = SA2.merge(rent_price, left_on = "SA2_NAME21", right_on = "SA2", how ='inner')

# plot the price map for residence
m2 = folium.Map(location=[-37.8136, 144.9631], tiles="Stamen Terrain", zoom_start=10)

folium.Choropleth(
    geo_data=geoJSON,
    data=rent_price,
    columns = ['SA2','per_room'],
    key_on='properties.SA2_NAME21',  
    fill_color='YlOrRd', 
    fill_opacity=0.7,
    line_opacity=0.2,
    legend_name='log of median rent per room price',
).add_to(m2)

m2.save('../plots/rent_price_pr_SA2.html')
